Rename CrossValidationItem to CrossValidationReporterItem
