+++ skore/src/skore/persistence/item/cross_validation_reporter_item.py
